requirements.txt
authlib>= 1.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
gunicorn==20.1.0
flask==3.0.3
Flask-Login==0.6.3
flask-session==0.8.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
python-dotenv==1.0.1
requests==2.31.0
urllib3==2.2.1
werkzeug==3.0.2
zipp==3.18.1